Volume, RichNode, SplitVolumeExample: moves function positionSlices from Volume to positionVolumeSlices in RichNode
